Attempt at integrating in C64 code
Still buggy, I think that the underflow handling is broken and
subtracting 129 as in the python code has not the same effect as
on the 6502 since we're not doing the subtraction in 16 bit.
I think the benefit of (e1+e2)-129 is that, if (e1+e2) overflows,
we get the overflowed 1 back in the exponent.